Refactor api app, organize code
