+++ api/routes/data.js
@@ -1,11 +1,12 @@
-import express from 'express'
+import { Router } from 'express'
 import dataController from '../controllers/dataController.js'
 import { isLoggedIn, isEnabled, isUploader } from '../lib/functions.js' // Helper functions
 
-const router = express.Router()
+const router = Router()
 
 const isValid = [isLoggedIn, isEnabled, isUploader]
 
+// routes for /data
 router.get('/', isValid, dataController.getData)
 router.get('/perUsers', isValid, dataController.getDataPerUsers)
 router.get('/perImages', isValid, dataController.getDataPerImages)
