+++ api/routes/users.js
@@ -1,11 +1,12 @@
-import express from 'express'
+import { Router } from 'express'
 import userController from '../controllers/userController.js'
 import { isLoggedIn, isEnabled, isAdmin } from '../lib/functions.js' // Helper functions
 
-const router = express.Router()
+const router = Router()
 
 const isValid = [isLoggedIn, isEnabled, isAdmin]
 
+// routes for /users
 router.get('/', isValid, userController.getUsers)
 router.post('/', isValid, userController.createUser)
